New version which incorporates a new ac-wrapper for a few bug fixes (#47772 and #41389)
Use a subroutine which will pull out the greatest AC_PREREQ() value in
a configure.in file in case there is more than one (see Bug #41389).
Also, change the gt statements to be 2.13 instead of 2.1 or AC_PREREQs fail
on seeing 2.12 for example (actually, don't fail, but use autoconf-2.5x which
is not correct).