autoconf-wrapper: export WANT_AUTOCONF to the full version
When exporting the WANT_AUTOCONF version based on what we picked, use
the full version instead of the short 2.1 or 2.5 values since they may
expand into an unrelated version (since we accept many versions as an
alias to 2.5).  Practically speaking, there shouldn't be a difference,
but this will make things a bit more clear.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>