autoconf-wrapper: slightly rework handling of 2.1 version alias
The 2.1 alias only ever expanded into the 2.13 version.  Every other
version uses the 2.5 alias, and at this point, seems extremely unlikely
that this will ever change (or that we'd add a new alias set).  Rework
the WANT_AUTOCONF checking logic to rewrite 2.1 to 2.13, and then accept
any non-2.13 version when 2.5 is used.  This will allow us to burn down
the alias logic entirely which will simplify the code nicely.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>